[Auto Generated] v0.0.354

+++ sdk/common/Paginator.go
@@ -1,12 +1,14 @@
 package common
 
+//Paginator holds paginator object
 type Paginator struct {
-	PageNo      int
+	PageNo      float64
 	HasNextPage bool
 	NextID      string
 	Next        func() (interface{}, error)
 }
 
+//NewPaginator returns paginator instance
 func NewPaginator(pageType string) *Paginator {
 	if pageType == "number" {
 		return &Paginator{
@@ -25,12 +27,14 @@ func NewPaginator(pageType string) *Paginator {
 	return &Paginator{}
 }
 
+//HasNext return has_next value
 func (p *Paginator) HasNext() bool {
 	return p.HasNextPage
 }
 
+//SetPaginator sets paginator instance with either next_id or page_no
 func (p *Paginator) SetPaginator(hasNext bool, pageNo int, nextID string) {
 	p.HasNextPage = hasNext
 	p.NextID = nextID
-	p.PageNo = pageNo
+	p.PageNo = float64(pageNo)
 }
